+++ spec/support/fixtures/query.json
@@ -1,0 +1,52 @@
+{
+  "query":{
+    "filtered":{
+      "query":{
+        "query_string":{
+          "analyze_wildcard":true,
+          "query":"memoryAddress:['A', '2E64930W'] OR memoryAddress:['B', '2E64932W'] OR memoryAddress:['C', '2E64934W'] OR memoryAddress:['D', '2E64936W']"
+        }
+      },
+      "filter":{
+        "bool":{
+          "must":[
+            {
+              "range":{
+                "timeStamp":{
+                  "gte":1441101600000,
+                  "lte":1441105200000
+                }
+              }
+            }
+          ],
+          "must_not":[
+
+          ]
+        }
+      }
+    }
+  },
+  "size":0,
+  "aggregations":{
+    "weight_chart":{
+      "date_histogram":{
+        "field":"timeStamp",
+        "interval":"10m",
+        "pre_zone":"+01:00",
+        "pre_zone_adjust_large_interval":true,
+        "min_doc_count":1,
+        "extended_bounds":{
+          "min":1441101600000,
+          "max":1441105200000
+        }
+      },
+      "aggregations":{
+        "weight":{
+          "avg":{
+            "field":"value"
+          }
+        }
+      }
+    }
+  }
+}
